+++ src/main/scala/com/foursquare/rogue/QueryClause.scala
@@ -13,9 +13,8 @@ object IndexBehavior extends Enumeration {
 }
 
 class QueryClause[V](val fieldName: String, conditions: (CondOps.Value, V)*) {
-  def extend(q: BasicDBObjectBuilder, signature: Boolean): BasicDBObjectBuilder = {
+  def extend(q: BasicDBObjectBuilder, signature: Boolean): Unit = {
     conditions foreach { case (op, v) => q.add(op.toString, if (signature) 0 else v) }
-    q
   }
   lazy val actualIndexBehavior = conditions.head._1 match {
     case CondOps.All | CondOps.In => IndexBehavior.Index
@@ -27,14 +26,22 @@ class QueryClause[V](val fieldName: String, conditions: (CondOps.Value, V)*) {
   def withExpectedIndexBehavior(b: IndexBehavior.Value) = new QueryClause(fieldName, conditions: _*) { override val expectedIndexBehavior = b }
 }
 
+class RawQueryClause(f: BasicDBObjectBuilder => Unit) extends QueryClause("raw") {
+  override def extend(q: BasicDBObjectBuilder, signature: Boolean): Unit = {
+    f(q)
+  }
+  override lazy val actualIndexBehavior = IndexBehavior.DocumentScan
+  override val expectedIndexBehavior = IndexBehavior.DocumentScan
+}
+
 class EmptyQueryClause[V](fieldName: String) extends QueryClause[V](fieldName) {
-  override def extend(q: BasicDBObjectBuilder, signature: Boolean): BasicDBObjectBuilder = new BasicDBObjectBuilder
+  override def extend(q: BasicDBObjectBuilder, signature: Boolean): Unit = {}
   override lazy val actualIndexBehavior = IndexBehavior.Index
   override def withExpectedIndexBehavior(b: IndexBehavior.Value) = new EmptyQueryClause[V](fieldName) { override val expectedIndexBehavior = b }
 }
 
 class EqClause[V](fieldName: String, value: V) extends QueryClause[V](fieldName) {
-  override def extend(q: BasicDBObjectBuilder, signature: Boolean): BasicDBObjectBuilder = {
+  override def extend(q: BasicDBObjectBuilder, signature: Boolean): Unit = {
     q.add(fieldName, if (signature) 0 else value)
   }
   override lazy val actualIndexBehavior = IndexBehavior.Index
@@ -42,7 +49,7 @@ class EqClause[V](fieldName: String, value: V) extends QueryClause[V](fieldName)
 }
 
 class WithinCircleClause[V](fieldName: String, lat: Double, lng: Double, radius: Double) extends QueryClause(fieldName) {
-  override def extend(q: BasicDBObjectBuilder, signature: Boolean): BasicDBObjectBuilder = {
+  override def extend(q: BasicDBObjectBuilder, signature: Boolean): Unit = {
     val value = if (signature) 0 else QueryHelpers.list(List(QueryHelpers.list(List(lat, lng)), radius))
     q.push("$within").add("$center", value).pop
   }
@@ -51,7 +58,7 @@ class WithinCircleClause[V](fieldName: String, lat: Double, lng: Double, radius:
 }
 
 class WithinBoxClause[V](fieldName: String, lat1: Double, lng1: Double, lat2: Double, lng2: Double) extends QueryClause(fieldName) {
-  override def extend(q: BasicDBObjectBuilder, signature: Boolean): BasicDBObjectBuilder = {
+  override def extend(q: BasicDBObjectBuilder, signature: Boolean): Unit = {
     val value = if (signature) 0 else {
       QueryHelpers.list(List(QueryHelpers.list(lat1, lng1), QueryHelpers.list(lat2, lng2)))
     }
@@ -62,14 +69,13 @@ class WithinBoxClause[V](fieldName: String, lat1: Double, lng1: Double, lat2: Do
 }
 
 class ModifyClause[V](val operator: ModOps.Value, fields: (String, V)*) {
-  def extend(q: BasicDBObjectBuilder): BasicDBObjectBuilder = {
+  def extend(q: BasicDBObjectBuilder): Unit = {
     fields foreach { case (name, value) => q.add(name, value) }
-    q
   }
 }
 
 class ModifyAddEachClause[V](fieldName: String, values: Traversable[V]) extends ModifyClause[V](ModOps.AddToSet) {
-  override def extend(q: BasicDBObjectBuilder): BasicDBObjectBuilder = {
+  override def extend(q: BasicDBObjectBuilder): Unit = {
     q.push(fieldName).add("$each", QueryHelpers.list(values)).pop
   }
 }